% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\docType{data}
\name{fix.50bin}
\alias{fix.50bin}
\title{Fixation proportion data for every 50 ms bin in the window from -1000 ms to 1000 ms relative to the target word onset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 29397 rows and 11 columns.
}
\usage{
fix.50bin
}
\description{
Data from Ito, Pickering & Corley (2018, JML)
}
\details{
\subsection{DATA STRUCTURE}{

Subject: subject ID

Trial: trial number

Bin: bin index

Time: time relative to the target word onset (ms)

allSample: the sum of all samples in the corresponding time bin

Count: right-eye sample count on the critical object

Condition: condition label (Targ=target, Eng=English competitor, Jap=Japanese competitor, Unr=unrelated)

Item: item ID

Lang: language group (L1=native English speakers, L2=native Japanese, non-native English speakers)

FixP: Fixation proportion

elogFix: empirical logit for each bin
}
}
\keyword{datasets}
