% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\docType{data}
\name{fix.binom.50bin.raw}
\alias{fix.binom.50bin.raw}
\title{Binomially coded fixation data for every 50 ms bin in the window from -1000 ms to 1000 ms relative to the target word onset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 29397 rows and 8 columns.
}
\usage{
fix.binom.50bin.raw
}
\description{
Data from Ito, Pickering & Corley (2018, JML)
This data set has not been modified.
}
\details{
\subsection{DATA STRUCTURE}{

Subject: subject ID

Trial: trial number

Bin: bin index

Time: time relative to the target word onset (ms)

Count: right-eye fixation on the critical object (1=fixated, 0=not fixated)

Condition: condition label (Targ=target, Eng=English competitor, Jap=Japanese competitor, Unr=unrelated)

Item: item ID

Lang: language group (L1=native English speakers, L2=native Japanese, non-native English speakers)
}
}
\keyword{datasets}
