% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\docType{data}
\name{fix.window.raw}
\alias{fix.window.raw}
\title{Fixation proportion data in the window from -800 ms to 0 ms relative to the target word onset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 717 rows and 8 columns.
}
\usage{
fix.window.raw
}
\description{
Data from Ito, Pickering & Corley (2018, JML)
This data set has not been modified.
}
\details{
\subsection{DATA STRUCTURE}{

Subject: subject ID

Trial: trial number

allSample: the sum of all samples in the corresponding time bin

Count: right-eye sample count on the critical object

FixP: Fixation proportion

Condition: condition label (Targ=target, Eng=English competitor, Jap=Japanese competitor, Unr=unrelated)

Item: item ID

Lang: language group (L1=native English speakers, L2=native Japanese, non-native English speakers)
}
}
\keyword{datasets}
