% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.VWP-package.R
\docType{package}
\name{stats.VWP-package}
\alias{stats.VWP}
\alias{stats.VWP-package}
\title{stats.VWP: Analysing visual world eye-tracking data in R}
\description{
This package contains data sets for the course "Analysing visual world eye-tracking data in R".
}
\author{
\strong{Maintainer}: Aine Ito \email{aine.ito@hu-berlin.de} (\href{https://orcid.org/0000-0003-4408-8801}{ORCID})

}
\keyword{internal}
